Remove SolverABC
With the new Builder/Designer architecture, this Solver base class no
longer has a place.

All the functionality of the base class has been moved to the
CodesSolver class, which is the only place the base class was really
being used.
