+++ tests/codes/test_interface.py
@@ -22,8 +22,8 @@
 from dataclasses import dataclass
 
 from bluemira.base.parameter_frame import Parameter
-from bluemira.base.solver import NoOpTask, RunMode
 from bluemira.codes import interface
+from bluemira.codes.interface import NoOpTask, RunMode
 from bluemira.codes.params import MappedParameterFrame, ParameterMapping
